backend/pyproject.toml
[project]
name = "rky-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"fastapi>=0.115.0",
	"passlib>=1.7.4",
	"python-jose>=3.3.0",
	"python-multipart>=0.0.9",
	"sqlmodel>=0.0.22",
	"uvicorn[standard]>=0.30.0",
]
